requirements.txt
asgiref==3.3.1
autopep8==1.5.4
count-files==1.5.0
Django==3.1.3
django-debug-toolbar==3.1.1
django-etc==1.2.0
django-extensions==3.0.9
django-hitcount==1.3.3
mypy==0.790
mypy-extensions==0.4.3
optimize-images==1.4.0
piexif==1.1.3
Pillow==8.0.1
pycodestyle==2.6.0
pydot==1.4.1
pyparsing==2.4.7
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
